Reafactor pasapi safe.go
