+++ pkg/cybr/api/requests/updatesafe.go
@@ -1,0 +1,9 @@
+package requests
+
+// UpdateSafe contains the body of the Update Safe function's request
+type UpdateSafe struct {
+	SafeName    string `json:"SafeName,omitempty"`
+	Description string `json:"Description,omitempty"`
+	OLACEnabled bool   `json:"OLACEnabled,omitempty"`
+	ManagingCPM string `json:"ManagingCPM,omitempty"`
+}
